impl play lines and SCL 8895
