+++ configs/8895/index.ts
@@ -1,0 +1,62 @@
+import { IConfig } from "../../lib/types"
+
+export const config: IConfig = {
+    freeSpin: {
+        conditions: [
+            { count: 2, multiply: 0, total: 5 },
+            { count: 3, multiply: 0, total: 10 },
+            { count: 4, multiply: 0, total: 15 },
+            { count: 5, multiply: 0, total: 20 },
+        ],
+        index: 12,
+    },
+    m: [
+        [1, 1, 1, 1, 1],
+        [0, 0, 0, 0, 0],
+        [2, 2, 2, 2, 2],
+        [0, 1, 2, 1, 0],
+        [2, 1, 0, 1, 2],
+        [0, 0, 1, 0, 0],
+        [2, 2, 1, 2, 2],
+        [1, 0, 0, 0, 1],
+        [1, 2, 2, 2, 1],
+        [1, 1, 0, 1, 1],
+        [1, 1, 2, 1, 1],
+        [2, 2, 2, 1, 0],
+        [0, 0, 0, 1, 2],
+        [2, 1, 0, 0, 0],
+        [0, 1, 2, 2, 2],
+        [0, 2, 0, 2, 0],
+        [2, 0, 2, 0, 2],
+        [1, 0, 1, 0, 1],
+        [1, 2, 1, 2, 1],
+        [0, 1, 0, 1, 0],
+        [2, 1, 2, 1, 2],
+    ],
+    p: [
+        [0, 0, 1, 10, 50],            // first
+        [0, 0, 5, 15, 20],         // second
+        [0, 0, 10, 20, 150],
+        [0, 0, 15, 40, 200],
+        [0, 0, 20, 50, 300],
+        [0, 0, 25, 100, 600],
+        [0, 0, 40, 200, 1000],
+        [0, 1, 50, 300, 1500],
+        [0, 5, 100, 400, 2500],
+        [0, 10, 150, 600, 5000],
+        [0, 20, 200, 1000, 5000],
+        [0, 0, 0, 0, 0],    // wild
+        [0, 0, 0, 0, 0],    // free
+    ],
+    r: 3,
+    w: [
+        [47, 38, 18, 26, 9, 7, 6, 20, 12, 11, 2, 8, 5],
+        [47, 38, 18, 26, 9, 7, 6, 20, 12, 11, 2, 5, 5],
+        [49, 39, 28, 26, 9, 6, 6, 13, 12, 6, 2, 5, 5],
+        [49, 41, 35, 30, 13, 6, 6, 6, 6, 4, 2, 5, 5],
+        [49, 41, 35, 30, 22, 11, 8, 2, 2, 2, 2, 2, 5],
+    ],
+    wild: {
+        index: 11,
+    }
+}
